feat: 포트폴리오 추가 & fix: 기타 개선사항 수정
