+++ next.config.ts
@@ -23,7 +23,7 @@ const nextConfig: NextConfig = {
   },
 
   experimental: {
-    optimizePackageImports: ['framer-motion', 'lucide-react'],
+    optimizePackageImports: ['framer-motion', 'react-icons'],
   },
 
   compiler: {
