+++ package.json
@@ -14,10 +14,10 @@
     "clsx": "^2.1.1",
     "framer-motion": "^12.23.1",
     "gray-matter": "^4.0.3",
-    "lucide-react": "^0.525.0",
     "next": "15.3.5",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-icons": "^5.5.0",
     "react-markdown": "^10.1.0",
     "rehype-highlight": "^7.0.2",
     "rehype-raw": "^7.0.0",
